requirements.txt
pybullet>=3.2.6
numpy>=1.23
Pillow>=10.0
imageio>=2.31
